Fixed regressions

+++ src/main/java/co/empathy/p01/model/Title.java
@@ -3,10 +3,10 @@ package co.empathy.p01.model;
 import java.util.List;
 
 public record Title(String id, String type, String primaryTitle, String originalTitle, Boolean isAdult,
-        Integer startYear, Integer endYear, Integer runtimeMinutes, List<String> genres, Float averageRating,
+        Integer startYear, Integer endYear, Integer runtimeMinutes, List<String> genres, Double averageRating,
         Long numVotes) {
 
-    public Title withRating(float averageRating, long numVotes) {
+    public Title withRating(double averageRating, long numVotes) {
         return new Title(id(), type(), primaryTitle(), originalTitle(), isAdult(), startYear(), endYear(),
                 runtimeMinutes(), genres(), averageRating, numVotes);
